Added bin/www to the git repo so that the server can actually start by downloading and running startup.pl.
Also, added config.bowerrc in an attempt to make bower install the needed components for the web interface in the public directory. Currently this does not work.

Also started using WebStorm to edit the project, so the files needed for the IDE are included in this commit.

Moreover, var's in the blog.spec.js file which were never mutated were changed to const.

 TODO: Fix config.bowerrc